[REFACTOR] rename serializer.py
